+++ Adapter/PlentymarketsAdapter/ResponseParser/Product/Image/ImageResponseParser.php
@@ -96,6 +96,8 @@ class ImageResponseParser implements ImageResponseParserInterface
             $image->setMediaIdentifier($media->getIdentifier());
             $image->setShopIdentifiers(array_filter($shopIdentifiers));
             $image->setPosition((int) $entry['position']);
+            $image->setName($name);
+            $image->setTranslations($media->getTranslations());
 
             return $image;
         } catch (Exception $exception) {