+++ Connector/TransferObject/Product/Image/Image.php
@@ -2,9 +2,11 @@
 
 namespace SystemConnector\TransferObject\Product\Image;
 
+use SystemConnector\TransferObject\TranslateableInterface;
 use SystemConnector\ValueObject\AbstractValueObject;
+use SystemConnector\ValueObject\Translation\Translation;
 
-class Image extends AbstractValueObject
+class Image extends AbstractValueObject implements TranslateableInterface
 {
     /**
      * @var string
@@ -20,6 +22,16 @@ class Image extends AbstractValueObject
      * @var int
      */
     private $position = 0;
+
+    /**
+     * @var string
+     */
+    private $name;
+
+    /**
+     * @var Translation[]
+     */
+    private $translations = [];
 
     /**
      * @return mixed
@@ -67,6 +79,38 @@ class Image extends AbstractValueObject
     public function setPosition($position)
     {
         $this->position = $position;
+    }
+
+    /**
+     * @return string
+     */
+    public function getName()
+    {
+        return $this->name;
+    }
+
+    /**
+     * @param string $name
+     */
+    public function setName($name)
+    {
+        $this->name = $name;
+    }
+
+    /**
+     * @return Translation[]
+     */
+    public function getTranslations()
+    {
+        return $this->translations;
+    }
+
+    /**
+     * @param Translation[] $translations
+     */
+    public function setTranslations(array $translations)
+    {
+        $this->translations = $translations;
     }
 
     /**